Created battery and added images
Created a battery class.  Resized the battery image to eliminate
unnecessary blank space on the sides.

Added a condition in the EntityManager collide loop which checks if the
entity is the player.  If so, it collides it with all other entities.
Then added a battery condition in the player's hitWith function to pick
up the battery.

Added a hasPower variable to the LaserGun.  It is false when
constructed, and becomes true when the player picks up a battery.

Changed the levels which are loaded to be those matching
"./lvl/level[num].json".  Removed check on enemies dead for testing of
the first level.

Added images.js which contains the enums IMAGE and ANIMS.  This is where
all images and animations will reside.  Renamed eye and eye-base to
turret and turretBase.  Added player sprites.